Struct eagre_asn1::types::Any

pub struct Any {
    // some fields omitted
}

Asn1 Any Type

Any encoded "Any" type will not be visible in the encoded bytes. It is a helper type, which should be used in case the actual type is not known yet.

Example

let any = Any::new("I am a random string".to_string()).unwrap();
let encoded = any.der_bytes().unwrap();
// Send to far away planet
let decoded = Any::der_from_bytes(encoded).unwrap();
assert_eq!("I am a random string", &decoded.resolve::<String>().unwrap());

Methods

impl Any

fn new<T: DER>(val: T) -> Result<Any>

Create a new Any object from a inner value

fn resolve<T: DER>(&self) -> Result<T>

Resolve the inner value of an Any object