Struct eagre_asn1::types::VideotexString

pub struct VideotexString(_);

Asn1 String Type

Currently restricted character sets are not enforced, so it is the callers job to check wether string contents are legal for the specific string type